[Update] banner image functionality
