+++ src/components/Marketing/AdBannersListing/index.js
@@ -28,7 +28,7 @@ export const AdBannersListing = (props) => {
       description: t('HOME_BANNER_DESCRIPTION', 'Add a banner to your home page and inform everyone about the latest news in your marketplace.'),
       icon: <HouseDoor />,
       info: t('RECOMMENDED_IMAGE_SIZE', 'Recommended size _size_').replace('_size_', '1440 x 300'),
-      aspectRatio: 1440 / 260,
+      aspectRatio: 1440 / 300,
       isSearhShow: false
     },
     {
@@ -37,7 +37,7 @@ export const AdBannersListing = (props) => {
       description: t('BUSINESS_LISTING_BANNER_DESCRIPTION', 'Add a banner inside your business listing to let everyone know about your latest promotions inside your businesses listed.'),
       icon: <ViewStacked />,
       info: t('RECOMMENDED_IMAGE_SIZE', 'Recommended size _size_').replace('_size_', '1440 x 300'),
-      aspectRatio: 1440 / 260,
+      aspectRatio: 1440 / 300,
       isSearhShow: false
     },
     {
@@ -46,9 +46,8 @@ export const AdBannersListing = (props) => {
       description: t('BUSINESS_PAGE_BANNER_DESCRIPTION', 'Add a banner inside the business page, and inform news and promotions specifically of the stores customers are ordering from.'),
       icon: <Shop />,
       info: t('RECOMMENDED_IMAGE_SIZE', 'Recommended size _size_').replace('_size_', '1440 x 300'),
-      aspectRatio: 1440 / 260,
-      isSearhShow: true,
-      isLink: true
+      aspectRatio: 1440 / 300,
+      isSearhShow: true
     },
     {
       key: 'app_home_page',
@@ -75,8 +74,7 @@ export const AdBannersListing = (props) => {
       icon: <Shop />,
       info: t('RECOMMENDED_IMAGE_SIZE', 'Recommended size _size_').replace('_size_', '300 x 300'),
       aspectRatio: 300 / 300,
-      isSearhShow: true,
-      isLink: true
+      isSearhShow: true
     }
   ]
 
@@ -139,7 +137,6 @@ export const AdBannersListing = (props) => {
             defaultPosition={openPositionDetail?.key}
             aspectRatio={openPositionDetail?.aspectRatio}
             isSearhShow={openPositionDetail?.isSearhShow}
-            isLink={openPositionDetail?.isLink}
           />
         </SideBar>
       )}
