+++ src/components/Marketing/BannerDetails/styles.js
@@ -70,33 +70,26 @@ export const ImagesContainer = styled.div`
   display: flex;
   flex-wrap: wrap;
 `
-export const BannerImageContainer = styled.div`
-  margin-bottom: 15px;
+export const BannerImage = styled.div`
+  cursor: pointer;
   ${({ active }) => active && css`
     border-radius: 8px;
     border: 1px solid ${props => props.theme.colors.primary};
   `}
-`
-export const BannerImage = styled.div`
   width: 130px;
   height: 130px;
-  margin: 10px;
+  margin: 10px 10px 15px 10px;
+  padding: 1px;
   border-radius: 8px;
   overflow: hidden;
-  cursor: -webkit-grab;
-  cursor: grab;
   box-shadow: 0px 1px 4px rgba(0, 0, 0, 0.1);
   position: relative;
 
-  img,
-  div {
+  img, div {
     width: 100%;
     border-radius: 8px;
     height: 100%;
     overflow: hidden;
-  };
-
-  img{
     object-fit: cover;
   }
 `
@@ -121,8 +114,8 @@ export const UploadImageIconContainer = styled.div`
 
   p {
     margin: 0px;
-    padding-top: 14px;
-    font-size: 12px;
+    padding-top: 10px;
+    font-size: 14px;
     color: ${props => props.theme.colors.lightGray};
     text-align: center;
   }
@@ -192,64 +185,6 @@ export const ActionSelectorWrapper = styled.div`
       }
       .dropdown-item:last-child {
         color: #E63757;
-      }
-    }
-  }
-`
-export const ActionsContainer = styled.div`
-`
-export const DropDownWrapper = styled.div`
-  ${props => props.theme?.rtl ? css`
-    margin-right: 5px;
-    ` : css`
-    margin-left: 5px;
-  `}
-  button {
-    display: flex;
-    background: transparent !important;
-    border: none;
-    padding: 5px;
-    border-radius: 8px;
-    svg {
-      color: ${props => props.theme.colors.headingColor};
-      font-size: 20px;
-    }
-    &:active,
-    &:focus {
-      border-color: unset !important;
-      box-shadow: none !important;
-    }
-    &:after {
-      display: none;
-    }
-    &:hover {
-      background: ${props => darken(0.04, props.theme.colors.secundary)} !important;
-    }
-    &:active {
-      background: ${props => darken(0.1, props.theme.colors.secundaryDarkContrast)} !important;
-    }
-  }
-
-  .show {
-    >div {
-      border: 1px solid ${props => props.theme.colors.borderColor};
-      box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.12);
-    }
-  }
-
-  > div {
-    > div {
-      border-radius: 8px;
-      .dropdown-item {
-        font-size: 12px;
-        color: ${props => props.theme.colors.headingColor};
-        padding: 7px 20px;
-        &:active {
-          background: ${props => darken(0.1, props.theme.colors.secundary)} !important;
-        }
-      }
-      .dropdown-item:last-child {
-        color: ${props => props.theme.colors.danger};
       }
     }
   }
